% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_list_job_templates}
\alias{mediaconvert_list_job_templates}
\title{Retrieve a JSON array of up to twenty of your job templates}
\usage{
mediaconvert_list_job_templates(Category, ListBy, MaxResults, NextToken,
  Order)
}
\arguments{
\item{Category}{Optionally, specify a job template category to limit responses to only job templates from that category.}

\item{ListBy}{Optional. When you request a list of job templates, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.}

\item{MaxResults}{Optional. Number of job templates, up to twenty, that will be returned at one time.}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of job templates.}

\item{Order}{When you request lists of resources, you can optionally specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.}
}
\description{
Retrieve a JSON array of up to twenty of your job templates. This will return the templates themselves, not just a list of them. To retrieve the next twenty templates, use the nextToken string returned with the array
}
\section{Request syntax}{
\preformatted{mediaconvert$list_job_templates(
  Category = "string",
  ListBy = "NAME"|"CREATION_DATE"|"SYSTEM",
  MaxResults = 123,
  NextToken = "string",
  Order = "ASCENDING"|"DESCENDING"
)
}
}

\keyword{internal}
