% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconvert_operations.R
\name{mediaconvert_list_tags_for_resource}
\alias{mediaconvert_list_tags_for_resource}
\title{Retrieve the tags for a MediaConvert resource}
\usage{
mediaconvert_list_tags_for_resource(Arn)
}
\arguments{
\item{Arn}{[required] The Amazon Resource Name (ARN) of the resource that you want to list tags for. To get the ARN, send a GET request with the resource name.}
}
\description{
Retrieve the tags for a MediaConvert resource.
}
\section{Request syntax}{
\preformatted{mediaconvert$list_tags_for_resource(
  Arn = "string"
)
}
}

\keyword{internal}
